refactor: :art: update file structure and import paths for consistency
