+++ src/components/layouts/components/navbar/Navbar.tsx
@@ -1,15 +1,12 @@
 import { useState } from 'react';
 
-import { Link } from '@tanstack/react-router';
 import { motion, useScroll, useMotionValueEvent } from 'framer-motion';
 
-import { CommandMenu } from '@/components/command-menu.tsx';
-import { Icons } from '@/components/icons.tsx';
-import { MobileNav } from '@/components/mobile-nav.tsx';
-import { ModeToggle } from '@/components/mode-toggle.tsx';
-import { SwitchLanguage } from '@/components/switch-language.tsx';
-import { Button } from '@/components/ui/button.tsx';
-import { useNavConfig } from '@/config/nav-config.tsx';
+import { MainNav } from '@/components/layouts/components/navbar/components/main-nav.tsx';
+import { MobileNav } from '@/components/layouts/components/navbar/components/mobile-nav.tsx';
+import { CommandMenu } from '@/components/shared/command-menu.tsx';
+import { ModeToggle } from '@/components/shared/mode-toggle.tsx';
+import { SwitchLanguage } from '@/components/shared/switch-language.tsx';
 
 export default function Navbar() {
   const [hidden, setHidden] = useState<boolean>(false);
@@ -50,37 +47,5 @@ export default function Navbar() {
         </div>
       </div>
     </motion.header>
-  );
-}
-
-function MainNav() {
-  const { mainNav } = useNavConfig();
-
-  return (
-    <div className='hidden md:flex lg:gap-3'>
-      <Link to='/' className='me-2 lg:me-6'>
-        <Icons.logo className='h-20 w-20 lg:h-24 lg:w-24 text-primary' />
-      </Link>
-      <nav className='flex items-center gap-1 xl:gap-3'>
-        {mainNav.map((link) => {
-          return (
-            <Link
-              key={link.label}
-              hash={link.hash}
-              activeOptions={{ exact: true, includeHash: true }}
-              activeProps={{ className: 'text-primary' }}
-              className='transition-colors font-bold'
-            >
-              <Button
-                variant='ghost'
-                className='sm:text-[0.8rem] lg:text-[1rem] font-Lato hover:text-primary lg:p-3'
-              >
-                {link.label}
-              </Button>
-            </Link>
-          );
-        })}
-      </nav>
-    </div>
   );
 }
