+++ src/components/shared/sectors.tsx
