+++ src/pages/home/Home.tsx
@@ -1,11 +1,11 @@
 import MainLayout from '@/components/layouts/main-layout.tsx';
+import Sectors from '@/components/shared/sectors.tsx';
 import BlurFade from '@/components/ui/blur-fade.tsx';
 
-import About from '@/pages/home/sections/About.tsx';
-import Contact from '@/pages/home/sections/Contact.tsx';
-import Hero from '@/pages/home/sections/Hero.tsx';
-import QualityPolicy from '@/pages/home/sections/QualityPolicy.tsx';
-import Sectors from '@/pages/home/sections/Sectors.tsx';
+import About from '@/pages/home/sections/about.tsx';
+import Contact from '@/pages/home/sections/contact.tsx';
+import Hero from '@/pages/home/sections/hero.tsx';
+import QualityPolicy from '@/pages/home/sections/quality-policy.tsx';
 
 function Home() {
   return (
