+++ src/types/nav.d.ts
@@ -2,7 +2,6 @@ import React from 'react';
 
 export interface NavItem {
   title?: string;
-  href?: string;
   hash?: string;
   disabled?: boolean;
   external?: boolean;
