+++ src/types/sector.d.ts
@@ -2,4 +2,5 @@ export interface Sector {
   title: string;
   image: string;
   description?: string;
+  href: string;
 }
